chore: reorganize plugin layout and cleanup tooling
