+++ archive/README.md
@@ -1,6 +1,6 @@
 # Archived plugins
 
 This directory keeps historical copies of legacy plugin bundles. The active
-runtime only loads plugins from `plugins_alpha/` or the optional
+runtime only loads plugins from `plugins/` or the optional
 `plugins_user/` tree. Legacy plugins moved here remain available for reference
 but are not imported automatically.
