+++ core/services/__init__.py
@@ -1,0 +1,4 @@
+"""Service-layer modules for Alpha Core."""
+from __future__ import annotations
+
+__all__: list[str] = []
