+++ docs/LICENSE-THIRD-PARTY.md
