+++ plugins/_template/__init__.py
