+++ plugins/reader/LICENSE
