+++ pyproject.toml
@@ -1,0 +1,10 @@
+[build-system]
+requires = ["setuptools>=65", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "tgc-alpha-core"
+version = "0.0.0"
+description = "Packaging stub for TGC Alpha Core."
+readme = "README.md"
+requires-python = ">=3.11"
