+++ tests/test_plugin_import_guard.py
@@ -6,7 +6,7 @@ import sys
 from pathlib import Path
 
 ROOT = Path(__file__).resolve().parents[1]
-PLUGDIR = ROOT / "plugins_alpha" / "violator"
+PLUGDIR = ROOT / "plugins" / "violator"
 PLUGDIR.mkdir(parents=True, exist_ok=True)
 (PLUGDIR / "__init__.py").write_text("", encoding="utf-8")
